renames services and moves things around
