+++ lib/sign_in/sign_in.rb
@@ -1,0 +1,2 @@
+module SignIn
+end
